+++ src/stella_vslam/optimize/pose_optimizer.h
@@ -39,7 +39,7 @@ public:
                                   std::vector<bool>& outlier_flags) const
         = 0;
 
-    virtual unsigned int optimize_matched_scores(const data::frame& frm, Mat44_t& optimized_pose, std::vector<bool>& outlier_flags, std::vector<double>& matched_scores) const = 0;
+    virtual unsigned int optimize_matched_scores(const data::frame& frm, Mat44_t& optimized_pose, std::vector<bool>& outlier_flags, std::vector<float>& matched_scores) const = 0;
 };
 
 } // namespace optimize
